requirements.txt
aioboto3==15.0.0
aiobotocore==2.23.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aioitertools==0.12.0
aiolimiter==1.2.1
aiosignal==1.3.2
airportsdata==20250224
alembic==1.15.1
annotated-types==0.7.0
anyio==4.9.0
anytree==2.12.1
appdirs==1.4.4
APScheduler==3.11.0
ascii_colors==0.11.4
astor==0.8.1
asyncer==0.0.8
attrs==25.3.0
autograd==1.7.0
backoff==2.2.1
beartype==0.18.5
blake3==1.0.4
boto3==1.38.27
botocore==1.38.27
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorlog==6.9.0
compressed-tensors==0.9.3
configparser==7.2.0
contourpy==1.3.1
cryptography==43.0.3
cupy-cuda12x==13.4.1
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.4.1
Deprecated==1.2.18
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.17.0
dotenv==0.9.9
dspy==2.6.27
dspy-ai==2.6.27
einops==0.8.1
email_validator==2.2.0
fast-graphrag==0.0.5
fastapi==0.115.12
fastapi-cli==0.0.7
fastapi-sso==0.16.0
fastrlock==0.8.3
filelock==3.18.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
future==1.0.0
gensim==4.3.3
gguf==0.16.2
git-filter-repo==2.47.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.35.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.20.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graspologic==3.4.1
graspologic-native==1.2.3
greenlet==3.1.1
grpc-google-iam-v1==0.14.2
grpcio==1.73.1
grpcio-status==1.71.2
gunicorn==22.0.0
h11==0.14.0
hf-xet==1.0.5
hnswlib==0.8.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
hyppo==0.4.0
idna==3.10
igraph==0.11.9
importlib_metadata==8.0.0
instructor==1.10.0
interegular==0.3.3
jieba==0.42.1
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
json_repair==0.39.1
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.49
langchain-openai==0.3.8
langchain-text-splitters==0.3.7
langsmith==0.3.19
lark==1.2.2
-e git+https://github.com/HKUDS/LightRAG.git@1c870454f1286ab42c1ad3d61925c17bce773dd8#egg=lightrag_hku
litellm==1.74.6
llguidance==0.7.19
llvmlite==0.44.0
lm-format-enforcer==0.10.11
magicattr==0.1.6
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.0
mdurl==0.1.2
milvus-lite==2.4.12
mistral_common==1.5.4
modelscope==1.25.0
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.2.0
multiprocess==0.70.16
mypy-extensions==1.0.0
-e git+https://github.com/gusye1234/nano-graphrag.git@0cea71bc5df33eee40a62c5312ed52bb6a9f2c54#egg=nano_graphrag
nano-vectordb==0.0.2
neo4j==5.25.0
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.4
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
ollama==0.4.8
openai==1.97.0
opencv-python-headless==4.11.0.86
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-semantic-conventions-ai==0.4.3
optuna==4.2.1
optype==0.12.2
orjson==3.10.16
outlines==0.1.11
outlines_core==0.1.26
packaging==24.2
pandas==2.2.3
partial-json-parser==0.2.1.1.post5
patsy==1.0.1
pillow==11.1.0
pipmaster==0.9.2
POT==0.9.5
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.21.1
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
py-cpuinfo==9.0.0
pyahocorasick==2.1.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycountry==24.6.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
pymilvus==2.5.6
PyMySQL==1.1.1
PyNaCl==1.5.0
pynndescent==0.5.13
pyparsing==3.2.3
pysbd==0.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.18
pytz==2025.2
pyuca==1.2
PyYAML==6.0.2
pyzmq==26.4.0
ragas==0.2.2
ray==2.43.0
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.3
rpds-py==0.24.0
rq==2.2.0
rsa==4.9.1
s3transfer==0.13.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.12.0
scipy-stubs==1.16.0.2
seaborn==0.13.2
sentencepiece==0.2.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
statsmodels==0.14.4
sympy==1.13.1
tenacity==9.0.0
texttable==1.7.0
threadpoolctl==3.6.0
tiktoken==0.8.0
tokenizers==0.21.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.66.5
transformers==4.52.3
triton==3.2.0
typer==0.15.2
typing-inspect==0.9.0
typing_extensions==4.13.0
tzdata==2025.2
tzlocal==5.3.1
ujson==5.10.0
umap-learn==0.5.6
urllib3==2.3.0
uvicorn==0.22.0
uvloop==0.21.0
vertexai==1.71.1
vllm==0.8.5
voyageai==0.3.3
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
wrapt==1.17.2
xformers==0.0.29.post2
xgrammar==0.1.18
xlsxwriter==3.2.5
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
